requirements.txt
numpy==2.3.3
pandas==2.3.3
scipy>=1.11
pillow==11.3.0
pytesseract>=0.3.10
opencv-python>=4.9
tqdm>=4.66
python-dotenv>=1.0
pyyaml>=6.0
google-genai==1.55.0
symspellpy==6.9.0
wordfreq>=3.1
jiwer>=3.0
textdistance>=4.6
nltk>=3.8
bert-score>=0.3.13
torch>=2.6
transformers>=4.41
scikit-learn>=1.4
matplotlib>=3.8
seaborn>=0.13
ultralytics>=8.2
streamlit>=1.52
ipython>=8.20
sentencepiece>=0.1.99
